requirement.txt
mediapipe
pyserial
opencv-python